Moved ChannelStats class to separate file.
